feat(auto-news-id): Generated when adding news

+++ Assignment/src/main/java/controller/admin/NewsController.java
@@ -27,6 +27,18 @@ public class NewsController extends HttpServlet {
         String action = req.getParameter("action");
         String id = req.getParameter("id");
 
+     // Thêm phần xử lý action "getLatestId"
+        if ("getLatestId".equals(action)) {
+            String categoryId = req.getParameter("categoryId");
+            int num = dao.getLatestIdWithCategory(categoryId);
+            String newId = num < 100 ? "0" + num : String.valueOf(num);
+            String latestId = categoryId + newId;
+            
+            resp.setContentType("application/json");
+            resp.getWriter().write("{\"latestId\": \"" + latestId + "\"}");
+            return;
+        }
+        
         if ("delete".equals(action) && id != null) {
             dao.delete(id);
             resp.sendRedirect(req.getContextPath() + "/admin/news");
@@ -54,12 +66,19 @@ public class NewsController extends HttpServlet {
             // hiển thị tất cả khi keyword rỗng
             list = dao.getAll();
         }
-
-
+        
         req.setAttribute("listNews", list);
 
         List<Category> categories = dao.getAllCate();
         req.setAttribute("categories", categories);
+        
+        // Auto ID
+        String catId = categories.getFirst().getId();
+        int num = dao.getLatestIdWithCategory(catId);
+        String latestId = num < 100 ? ( num < 10 ? "00" + num : "0" + num ) : String.valueOf(num);
+        latestId = catId + latestId;
+        
+        req.setAttribute("latestId", latestId);
         req.setAttribute("searchBy", searchBy);
         req.setAttribute("keyword", keyword);
         req.getRequestDispatcher("/admin/news.jsp").forward(req, resp);
